requirements.txt
coverage
mock
nose
-e git://github.com/lacostej/nosyd.git#egg=nosyd
pyyaml
boto
